+++ src/main/java/jp/co/axa/apidemo/controllers/Paths.java
@@ -1,0 +1,9 @@
+package jp.co.axa.apidemo.controllers;
+
+public class Paths {
+    
+    public static final String EMPLOYEES = "employees";
+    public static final String ID = "id";
+    public static final String V1 = "v1";
+    public static final String __ = "/";
+}
