+++ src/main/resources/application.properties
@@ -1,1 +1,3 @@
 spring.h2.console.enabled=true
+server.servlet.contextPath=/api
+
